in the book entity class year was change from Long type to int primitive type

+++ bookstore/src/main/java/com/plakolli/bookstore/domain/Book.java
@@ -19,7 +19,7 @@ public class Book {
 	
 	public Book() {} // constructor without parameters.
 
-	public Book(String title, String author, String isbn, Long year, Float price) {
+	public Book(String title, String author, String isbn, int year, Float price) {
 		this.title = title;
 		this.author = author;
 		this.isbn = isbn;
